requirements.txt
tweepy==4.14.0
python-telegram-bot==20.6
aiohttp==3.8.6
asyncio==3.4.3